Changed Double to BigDecimal to allow for more fine-grained flooring

+++ src/main/java/com/sysunite/rws/deflecties/GpsNavigationResult.java
@@ -1,11 +1,13 @@
 package com.sysunite.rws.deflecties;
 
+import java.math.BigDecimal;
+
 /**
  * @author Mohamad Alamili <mohamad@sysunite.com>
  */
 public class GpsNavigationResult {
   public Integer failureCause = 0;
-  public Double latitude;
-  public Double longitude;
-  public Double height;
+  public BigDecimal latitude;
+  public BigDecimal longitude;
+  public BigDecimal height;
 }
