+++ src/main/java/com/sysunite/rws/deflecties/SD.java
@@ -1,15 +1,17 @@
 package com.sysunite.rws.deflecties;
 
+import java.math.BigDecimal;
+
 /**
  * @author Mohamad Alamili <mohamad@sysunite.com>
  */
 public class SD {
   public String id;
   public Integer type;
-  public Double relativeGain;
-  public Double absoluteGain;
+  public BigDecimal relativeGain;
+  public BigDecimal absoluteGain;
 
-  public SD(String id, Integer type, Double r, Double a) {
+  public SD(String id, Integer type, BigDecimal r, BigDecimal a) {
     this.id = id;
     this.type = type;
     relativeGain = r;
